pyproject.toml
[project]
name = "scholarship"
version = "0.1.0"
description = "Scholarship eligibility RAG assistant for India (Databricks + Sarvam)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  # Align with Databricks Apps template (stable with Gradio 4.44 + huggingface_hub).
  "pandas~=2.2.3",
  "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
# Databricks Vector Search SDK (replaces faiss-cpu + sentence-transformers).
vectorsearch = [
  "databricks-vectorsearch",
]
# Gradio UI for Databricks Apps (`python app/main.py`).
# Pin huggingface_hub with Gradio 4.44 — matches Databricks template.
app = [
  "gradio~=4.44.0",
  # Must match gradio 4.44.x wheel metadata; newer client breaks api_info JSON schema parsing.
  "gradio-client==1.3.0",
  "huggingface-hub~=0.35.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
markers = [
  "network: tests that call external HTTP (opt-in)",
]
